break up lints.rs into separate files
