+++ components/plugins/lib.rs
@@ -44,11 +44,11 @@ pub fn plugin_registrar(reg: &mut Registry) {
     reg.register_syntax_extension(intern("dom_struct"), Modifier(box jstraceable::expand_dom_struct));
     reg.register_syntax_extension(intern("jstraceable"), Decorator(box jstraceable::expand_jstraceable));
     reg.register_syntax_extension(intern("_generate_reflector"), Decorator(box reflector::expand_reflector));
-    reg.register_lint_pass(box lints::TransmutePass as LintPassObject);
-    reg.register_lint_pass(box lints::UnrootedPass as LintPassObject);
-    reg.register_lint_pass(box lints::PrivatizePass as LintPassObject);
-    reg.register_lint_pass(box lints::InheritancePass as LintPassObject);
-    reg.register_lint_pass(box lints::StrToStringPass as LintPassObject);
+    reg.register_lint_pass(box lints::transmute_type::TransmutePass as LintPassObject);
+    reg.register_lint_pass(box lints::unrooted_must_root::UnrootedPass as LintPassObject);
+    reg.register_lint_pass(box lints::privatize::PrivatizePass as LintPassObject);
+    reg.register_lint_pass(box lints::inheritance_integrity::InheritancePass as LintPassObject);
+    reg.register_lint_pass(box lints::str_to_string::StrToStringPass as LintPassObject);
 }
