+++ components/plugins/lints/mod.rs
@@ -1,0 +1,9 @@
+/* This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this
+ * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
+
+ pub mod inheritance_integrity;
+ pub mod privatize;
+ pub mod str_to_string;
+ pub mod transmute_type;
+ pub mod unrooted_must_root;
